Moved compilers into their own namespace/folders
