+++ SassAndCoffee.Core/Compilers/ISimpleFileCompiler.cs
@@ -1,4 +1,4 @@
-namespace SassAndCoffee.Core
+namespace SassAndCoffee.Core.Compilers
 {
     public interface ISimpleFileCompiler
     {
