+++ SassAndCoffee.Core/ICompilerHost.cs
@@ -1,5 +1,7 @@
 namespace SassAndCoffee.Core
 {
+    using SassAndCoffee.Core.Compilers;
+
     public interface ICompilerHost
     {
         ISimpleFileCompiler MapPathToCompiler(string physicalPath);
